+++ mastracode/src/constants.ts
@@ -1,0 +1,2 @@
+// Default OM model - using gemini-2.5-flash for efficiency
+export const DEFAULT_OM_MODEL_ID = process.env.DEFAULT_OM_MODEL_ID ?? "google/gemini-2.5-flash"
